refactor: move model registry to lib
