+++ components/Inputs/InputChat/index.tsx
@@ -4,8 +4,8 @@ import { Send, Square } from "lucide-react";
 import { type FC, useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { getModelConfig } from "@/constants/models";
 import { canSendSelectedModel, getSelectedModelError } from "@/lib/chat/config";
+import { getModelConfig } from "@/lib/models";
 import { cn } from "@/lib/utils";
 import { useConfig, useUIActions } from "@/store";
 import { ReasoningSelector } from "./ReasoningSelector";
