+++ lib/models/google.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { getModelConfig, getReasoningFields } from "./models";
+import { getModelConfig, getReasoningFields } from "./index";
 
 describe("getReasoningFields", () => {
   it("asks Google models to include returned thoughts", () => {
